requirements.txt
nemo_toolkit
fiddle
cloudpickle
kaldialign
nv-one-logger-core
nv-one-logger-pytorch-lightning-integration
nv-one-logger-training-telemetry
fastapi
uvicorn
python-multipart
pydantic
torch
numpy
hydra-core
lightning<2.6
lhotse
einops
transformers
sentencepiece
pandas
jiwer
librosa
pyannote-audio
webdataset
datasets
editdistance
IPython
pydub
ffmpeg-python
soundfile
cuda-python
